Hack to fix hang with V-Amp Pro (should be same as V-Amp 2).
Renamed original ReadSysex to ReadSysex_ and replaced with workaround. Seems
like I get GIO callback only on events initiated on V-AMP and not for SysEx
responses (thus hanging forever on identify or any subsequent SysEx
request). I only get 6080 bytes for all presets. It is crude but works for
me, allocating 8192 to be on the safe side.